BIP62: Make OP_0 a validly encoded signature
Previously BIP62 did not provide a compact way to delibrately encode an
invalid signature. For example in BIP19 if m != n with this change you
can provide compact OP_0's in the scriptSig rather than lengthy
DER-encoded signatures.

Note that we may want to further expand on this change in the future by
saying that only OP_0 is a "valid" invalid signature; BIP19 even with
this change is inherently malleable as the invalid signatures can be any
validly encoded DER signature.